+++ src/controllers/InstallController.php
@@ -21,6 +21,7 @@ use craft\models\Site;
 use craft\web\assets\installer\InstallerAsset;
 use craft\web\Controller;
 use yii\base\Exception;
+use yii\base\InvalidConfigException;
 use yii\base\Response;
 use yii\web\BadRequestHttpException;
 
@@ -69,9 +70,9 @@ class InstallController extends Controller
 
         // Can we establish a DB connection?
         try {
-            Craft::$app->getDb()->open();
+            $this->_checkDbConfig();
             $showDbScreen = false;
-        } catch (DbConnectException $e) {
+        } catch (InvalidConfigException $e) {
             // Can we control the settings?
             if ($this->_canControlDbConfig()) {
                 $showDbScreen = true;
@@ -307,6 +308,24 @@ class InstallController extends Controller
         }
 
         return $this->asJson(['success' => $success]);
+    }
+
+    /**
+     * @throws InvalidConfigException
+     */
+    private function _checkDbConfig(): void
+    {
+        // If no database is set yet, definitely show it
+        if (!Craft::$app->getConfig()->getDb()->database) {
+            throw new InvalidConfigException('No database has been selected yet.');
+        }
+
+        // Can we establish a DB connection?
+        try {
+            Craft::$app->getDb()->open();
+        } catch (DbConnectException $e) {
+            throw new InvalidConfigException('A database connection couldn’t be established.', 0, $e);
+        }
     }
 
     /**